requirements.txt
torchvision==0.7.0
torch==1.6.0
pandas==1.0.5
numpy==1.19.3
scikit_image==0.17.2
opencv_contrib_python==4.5.1.48
Pillow==8.2.0
skimage==0.0
